GlobalSequence and LocalSequence Gotoh are implemented

+++ src/bioinfo/alignment/SequenceAlignment.java
@@ -33,7 +33,7 @@ public class SequenceAlignment implements Alignment {
 	/**
 	 * The score of the Alignment
 	 */
-	private final double score;
+	private final int score;
 	/**
 	 * The length of the Alignment
 	 */
@@ -46,7 +46,7 @@ public class SequenceAlignment implements Alignment {
 	 * @param rows the two rows of the alignment
 	 * @param score the score of the Alignment
 	 */
-	public SequenceAlignment(Sequence seq1, Sequence seq2, char[][] rows, double score) {
+	public SequenceAlignment(Sequence seq1, Sequence seq2, char[][] rows, int score) {
 		this.seq1 = seq1;
 		this.seq2 = seq2;
 		this.rows = rows;
@@ -64,7 +64,7 @@ public class SequenceAlignment implements Alignment {
 	 * @param score the score of the Alignment
 	 */
 	public SequenceAlignment
-		(Sequence seq1, Sequence seq2, char[] row1, char[] row2, double score)
+		(Sequence seq1, Sequence seq2, char[] row1, char[] row2, int score)
 	{
 		this.seq1 = seq1;
 		this.seq2 = seq2;
@@ -86,7 +86,7 @@ public class SequenceAlignment implements Alignment {
 	 * @param score the score of the Alignment
 	 */
 	public SequenceAlignment
-		(Sequence seq1, Sequence seq2, String row1, String row2, double score)
+		(Sequence seq1, Sequence seq2, String row1, String row2, int score)
 	{
 		this.seq1 = seq1;
 		this.seq2 = seq2;
@@ -165,7 +165,7 @@ public class SequenceAlignment implements Alignment {
 		return length;
 	}
 	
-	public double getScore() {
+	public int getScore() {
 		return score;
 	}
